Split _test_functions up
